#9626 Move v2 package to main package
