+++ platform-sdk/swirlds-logging/src/main/java/com/swirlds/logging/extensions/event/SimpleLogMessage.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.logging.v2.extensions.event;
+package com.swirlds.logging.extensions.event;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
