+++ platform-sdk/swirlds-logging/src/main/java/com/swirlds/logging/extensions/event/package-info.java
@@ -2,4 +2,4 @@
  * This package contains the API for the logging events. This is the used as the basic data model of the logging API and
  * must be used by any extension to handle or provide logging events.
  */
-package com.swirlds.logging.v2.extensions.event;
+package com.swirlds.logging.extensions.event;
