+++ platform-sdk/swirlds-logging/src/main/java/com/swirlds/logging/extensions/package-info.java
@@ -2,4 +2,4 @@
  * This package (and all subpackages) contains the extension API of the logging framework. That API can be used to
  * create custom providers or handlers.
  */
-package com.swirlds.logging.v2.extensions;
+package com.swirlds.logging.extensions;
