--- platform-sdk/swirlds-logging/src/main/java/com/swirlds/logging/v2/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains the private API of the logging framework.
- */
-package com.swirlds.logging.v2;
