+++ platform-sdk/swirlds-logging/src/test/java/com/swirlds/logging/v2/LevelTest.java
@@ -20,6 +20,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.swirlds.logging.Level;
 import org.junit.jupiter.api.Test;
 
 class LevelTest {
