+++ platform-sdk/swirlds-logging/src/test/java/com/swirlds/logging/v2/MarkerTest.java
@@ -23,6 +23,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import com.swirlds.logging.Marker;
 import org.junit.jupiter.api.Test;
 
 public class MarkerTest {
