+++ platform-sdk/swirlds-logging/src/test/java/com/swirlds/logging/v2/emergency/EmergencyLoggerTest.java
@@ -23,11 +23,11 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.swirlds.base.test.fixtures.io.SystemErrProvider;
 import com.swirlds.base.test.fixtures.io.WithSystemError;
 import com.swirlds.base.test.fixtures.io.WithSystemOut;
-import com.swirlds.logging.v2.Level;
-import com.swirlds.logging.v2.Marker;
-import com.swirlds.logging.v2.extensions.event.LogEvent;
-import com.swirlds.logging.v2.extensions.event.LogMessage;
-import com.swirlds.logging.v2.internal.emergency.EmergencyLoggerImpl;
+import com.swirlds.logging.Level;
+import com.swirlds.logging.Marker;
+import com.swirlds.logging.extensions.event.LogEvent;
+import com.swirlds.logging.extensions.event.LogMessage;
+import com.swirlds.logging.internal.emergency.EmergencyLoggerImpl;
 import jakarta.inject.Inject;
 import java.time.Instant;
 import java.util.List;
